vehicle type crud operations

+++ TaxiAppsWebAPICore/TaxiAppsWebAPICore/Models/VehicleList.cs
@@ -6,7 +6,7 @@ namespace TaxiAppsWebAPICore
     public class VehicleTypeList
     {
         [JsonProperty("id")]
-        public int Id { set; get; }
+        public long Id { set; get; }
         [JsonProperty("name")]
         public string Name { set; get; }
 
@@ -20,7 +20,7 @@ namespace TaxiAppsWebAPICore
     public class VehicleTypeInfo
     {
         [JsonProperty("id")]
-        public int Id { set; get; }
+        public long Id { set; get; }
         [JsonProperty("name")]
         public string Name { set; get; }
